pyproject.toml
[project]
name = "a2a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]>=0.3.4",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
